Drop automagical repo initialization
bup had a convenience feature that commands automagically initialized a
repo in ~/.bup if it didn't exist and no other BUP_DIR was given.

This had the odd effect that when one forgot to give the BUP_DIR a bup
repo was initialized in ~/.bup even tough only a browsing command was
used.

This patch drops this behaviour for all commands.